Moved job_submit into the vcstools module
