+++ setup.py
@@ -52,17 +52,18 @@ setup(name="mwa_vcstools",
       python_requires='>=3.6',
       install_requires=reqs,
       scripts=['scripts/checks.py', 'scripts/calibrate_vcs.py', 'scripts/create_psrfits.sh',
-               'scripts/find_pulsar_in_obs.py', 'scripts/plot_BPcal_128T.py',
+               'scripts/find_pulsar_in_obs.py', 'scripts/RVM_fit.py',
                'scripts/process_vcs.py', 'scripts/recombine.py', 'scripts/rename_corr_output.py',
                'scripts/reorder_chans.py', 'scripts/rts2ao.py', 'scripts/untar.sh',
                'scripts/cleanup.py', 'scripts/create_ics_psrfits.py', 'scripts/rm_synthesis.py',
                'scripts/splice.sh', 'scripts/auto_plot.bash', 'scripts/splice_wrapper.py',
-               'scripts/RVM_fit.py',
+               # plotting scripts
+               'scripts/plotting/plotPolarTileBeam.py', 'scripts/plotting/plotFlatTileBeam.py',
+               'scripts/plotting/plotTiedArrayBeam.py', 'scripts/plotting/plotSkyMap.py',
+               'scripts/plotting/plot_BPcal_128T.py',
                'database/submit_to_database.py', 'database/database_vcs.py',
                'utils/zapchan.py', 'utils/calc_ephem.py', 'utils/check_disk_usage.sh',
-               'utils/check_quota.sh', 'utils/mdir.py', 'utils/mwa_metadb_utils.py',
-               'utils/job_submit.py', 'utils/plotFlatTileBeam.py', 'utils/plotPolarTileBeam.py',
-               'utils/plotTiedArrayBeam.py', 'utils/aocal.py', "utils/stickel.py",
+               'utils/check_quota.sh', 'utils/mdir.py', 'utils/mwa_metadb_utils.py','utils/aocal.py', "utils/stickel.py",
                'utils/config_vcs.py', 'utils/sn_flux_est.py', 'utils/prof_utils.py', 'utils/rm.py',
                'version.py'],
       setup_requires=['pytest-runner'],
